Consistent creation of Timelines with empty Segments
The previous behavior for initializing Timelines with empty Segments
was inconsistent with the behavior when adding one segment at a time.
Initializing raised an exception if any Segment was empty, but adding
empty segments simply did nothing.  With this change, initialization
now adds non-empty segments and ignores empty ones, resulting in a
consistent behavior with the case of adding Segments.

Added several tests to demonstrate that the behavior is now consistent.