fix(SlidingWindowFeature): fix PNG render with one feature
`plt.subplots()` when called without explicitly setting the `squeeze` argument
to False will try to squeeze out extra dimensions from the returned `axes`
ndarray. Currently the code relies on this behaviour to get a 1-dimensional
array when we pass in a hardcoded value of 1 for `ncols`, but does not count on
this squeezing out *all* dimensions and returning a single `Axes` object when
the (non-hardcoded) `nrows` is also 1.

To fix this we set the `squeeze` argument to False and manually call
`.flatten()` on the returned ndarray, to ensure we always get an exactly 1D
array of Axes out.